+++ src/mpkg/mpkg_pacidx.c
@@ -301,7 +301,7 @@ static void load_index(CTX)
 
 	if((ret = sys_read(fd, rest, need)) < 0)
 		fail("read", NULL, ret);
-	if(ret < need)
+	if(ret < (int)need)
 		fail("incomplete read", NULL, 0);
 }
